+++ CHANGELOG.md
@@ -5,12 +5,12 @@ All notable changes to "sower-tools" will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.12.15] - 2024-11-21
+## [0.13.0] - 2024-11-21
 
 ### Added
 
 -   Add encrypt/decrypt text.
--   Add configuration for encrypted/decrypted password.
+-   Add configuration for encrypted/decrypted text.
 
 ## [0.12.14] - 2024-11-18